가격과 수량 로그에 ToFixedString을 도입하여 더욱 정확한 소수점 정밀도로 로깅하고 과학적 표기법이 등장하지 않도록 수정

+++ Sources/Cpp/Engines/SymbolInfo.cpp
@@ -98,7 +98,7 @@ string SymbolInfo::GetExchangeInfoPath() const { return exchange_info_path_; }
 double SymbolInfo::GetPriceStep() const { return price_step_; }
 int SymbolInfo::GetPricePrecision() const { return price_precision_; }
 double SymbolInfo::GetQtyStep() const { return qty_step_; }
-size_t SymbolInfo::GetQtyPrecision() const { return qty_precision_; }
+int SymbolInfo::GetQtyPrecision() const { return qty_precision_; }
 double SymbolInfo::GetLimitMaxQty() const { return limit_max_qty_; }
 double SymbolInfo::GetLimitMinQty() const { return limit_min_qty_; }
 double SymbolInfo::GetMarketMaxQty() const { return market_max_qty_; }
